+++ backend/src/main/java/softeer/wantcar/cartalog/estimate/dao/EstimateDao.java
@@ -1,4 +1,4 @@
-package softeer.wantcar.cartalog.estimate.dto;
+package softeer.wantcar.cartalog.estimate.dao;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -7,7 +7,7 @@ import java.util.List;
 
 @Getter
 @Builder
-public class EstimateSaveDto {
+public class EstimateDao {
     private Long detailTrimId;
     private Long trimExteriorColorId;
     private Long trimInteriorColorId;
