+++ backend/src/main/java/softeer/wantcar/cartalog/estimate/repository/EstimateCommandRepository.java
@@ -1,7 +1,7 @@
 package softeer.wantcar.cartalog.estimate.repository;
 
-import softeer.wantcar.cartalog.estimate.dto.EstimateSaveDto;
+import softeer.wantcar.cartalog.estimate.dao.EstimateDao;
 
 public interface EstimateCommandRepository {
-    void save(EstimateSaveDto estimateSaveDto);
+    void save(EstimateDao estimateDao);
 }
