+++ backend/src/main/java/softeer/wantcar/cartalog/estimate/repository/EstimateQueryRepositoryImpl.java
@@ -12,6 +12,7 @@ import softeer.wantcar.cartalog.estimate.repository.dto.EstimateOptionListDto;
 import softeer.wantcar.cartalog.global.utils.RowMapperUtils;
 import softeer.wantcar.cartalog.estimate.dto.EstimateRequestDto;
 import softeer.wantcar.cartalog.estimate.dto.EstimateSaveDto;
+import softeer.wantcar.cartalog.estimate.dao.EstimateDao;
 
 import java.util.List;
