+++ backend/src/test/java/softeer/wantcar/cartalog/estimate/repository/EstimateCommandRepositoryImplTest.java
@@ -8,7 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.test.context.jdbc.Sql;
-import softeer.wantcar.cartalog.estimate.dto.EstimateSaveDto;
+import softeer.wantcar.cartalog.estimate.dao.EstimateDao;
 
 import java.util.List;
 
@@ -36,7 +36,7 @@ class EstimateCommandRepositoryImplTest {
         @DisplayName("생성 테스트")
         void createTest() {
             //given
-            EstimateSaveDto request = EstimateSaveDto.builder()
+            EstimateDao request = EstimateDao.builder()
                     .detailTrimId(9L)
                     .trimInteriorColorId(7L)
                     .trimExteriorColorId(7L)
